Cache articles with analysed title in realtime database

+++ app/src/main/java/com/csc306/coursework/model/LikabilityDTO.kt
@@ -1,3 +1,3 @@
 package com.csc306.coursework.model
 
-class LikabilityDTO (val keyword: String, val likabilityFactor: Float)
+class LikabilityDTO (val keyword: String, val likabilityFactor: Double)
